Updated width and height output for filename
Added 'dimensions' variable to determine if the aspect ratio of the icon is rectangular or square and pass either both dimensions (WxH) if rectangular, or one dimension (H) if square. This variable is used in the filename to give the developer the relevant DP or PT layout dimensions of the image as efficiently as possible.